% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buildings.R
\name{download_and_uncompress_buildings_data}
\alias{download_and_uncompress_buildings_data}
\title{Download and Uncompress Google Open Buildings Data}
\usage{
download_and_uncompress_buildings_data(
  region_code,
  output_dir,
  delete_compressed
)
}
\arguments{
\item{region_code}{A character string representing the region code.}

\item{output_dir}{The directory where the data will be saved.}

\item{delete_compressed}{Logical. Whether to delete the compressed file after uncompressing (default is \code{TRUE}).}
}
\value{
None. Downloads data to the specified \code{output_dir}.
}
\description{
This function downloads and uncompresses building data for a specific region code from Google Open Buildings.
}
\examples{
\dontrun{
download_and_uncompress_buildings_data("123", "gob_raw_data/")
}
}
