% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buildings.R
\name{download_buildings}
\alias{download_buildings}
\title{Download Building Data from a Specified Provider}
\usage{
download_buildings(
  centroid,
  provider = "google",
  output_dir = "gob_raw_data/",
  output_file = "osm_buildings.geojson",
  delete_compressed = TRUE
)
}
\arguments{
\item{centroid}{A numeric vector representing the centroid coordinates in the format \code{c(lon, lat)}.}

\item{provider}{The data provider to use. Options are \code{"google"} or \code{"osm"} (default is \code{"google"}).}

\item{output_dir}{The directory where data will be saved for Google data (default is \code{"gob_raw_data/"}).}

\item{output_file}{The file path where OSM data should be saved (default is \code{"osm_buildings.geojson"}).}

\item{delete_compressed}{Logical. Whether to delete the compressed file after uncompressing for Google data (default is \code{TRUE}).}
}
\value{
A \code{sf} object containing all building geometries within the specified area.
}
\description{
This function downloads building data from either Google Open Buildings or OpenStreetMap (OSM)
using a centroid point (longitude, latitude) and saves it to the specified directory or file.
}
\examples{
\dontrun{
download_buildings(c(4.35, 50.85), provider = "google")
download_buildings(c(4.35, 50.85), provider = "osm", output_file = "buildings.geojson")
}
}
