% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buildings.R
\name{download_google_open_buildings}
\alias{download_google_open_buildings}
\title{Download Google Open Buildings Data}
\usage{
download_google_open_buildings(
  centroid,
  output_dir = "gob_raw_data/",
  delete_compressed = TRUE
)
}
\arguments{
\item{centroid}{A numeric vector representing the centroid coordinates in the format \code{c(lon, lat)}.}

\item{output_dir}{The directory where data will be saved (default is \code{"gob_raw_data/"}).}

\item{delete_compressed}{Logical. Whether to delete the compressed file after uncompressing (default is \code{TRUE}).}
}
\value{
A \code{sf} object containing all building geometries within the specified region.
}
\description{
This function downloads building data from Google Open Buildings using a centroid point and saves the data to the specified directory.
}
\examples{
\dontrun{
download_google_open_buildings(c(4.35, 50.85))
}
}
