% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buildings.R
\name{get_open_buildings_region_code}
\alias{get_open_buildings_region_code}
\title{Get the Google Open Buildings Region Code}
\usage{
get_open_buildings_region_code(centroid)
}
\arguments{
\item{centroid}{A numeric vector representing the centroid coordinates in the format \code{c(lon, lat)}.}
}
\value{
A character string representing the S2 region code.
}
\description{
This function calculates the S2 region code for a given centroid.
}
\examples{
get_open_buildings_region_code(c(4.35, 50.85))
}
